changed feature array to Float64

+++ src/SortSpikes.jl
@@ -17,7 +17,7 @@ type Sorting{S<:SpikeDetection,C<:Cluster,A<:Alignment,F<:Feature}
     electrode::Array{Int64,1}
     neuronnum::Array{Int64,1}
     numSpikes::Int64
-    waveforms::Array{Array{Int64,1},1}
+    waveforms::Array{Array{Float64,1},1}
     sigend::Array{Int64,1}
     index::Int64
     p_temp::Array{Int64,1}
@@ -44,7 +44,7 @@ function Sorting(s::SpikeDetection,c::Cluster,a::Alignment,f::Feature)
       
     Sorting(s,c,a,f,
             zeros(Int64,signal_length),zeros(Int64,500),zeros(Int64,500),2,
-            [zeros(Int64,wavelength) for j=1:10], 
+            [zeros(Float64,wavelength) for j=1:10], 
             zeros(Int64,window+window_half),0,zeros(Int64,window*2),zeros(Float64,featurelength))   
 end
